moved testSubmoduleUpdateWithError to GitAPITestUpdateCliGit
