+++ src/test/java/org/jenkinsci/plugins/gitclient/GitAPITestUpdate.java
@@ -893,7 +893,6 @@ public abstract class GitAPITestUpdate {
         }
     }
 
-
     /* Shows the JGit submodule update is broken now that tests/getSubmodule includes a renamed submodule */
     @Test
     public void testGetSubmodules() throws Exception {
@@ -917,22 +916,6 @@ public abstract class GitAPITestUpdate {
             assertTrue("modules/sshkeys does not exist", w.exists("modules/sshkeys"));
         }
         assertFixSubmoduleUrlsThrows();
-    }
-
-    @NotImplementedInJGit
-    @Test
-    public void testSubmoduleUpdateWithError() throws Exception {
-        w.git.clone_().url(localMirror()).execute();
-        w.git.checkout().ref("origin/tests/getSubmodules").execute();
-        w.rm("modules/ntp");
-        w.touch("modules/ntp", "file that interferes with ntp submodule folder");
-
-        try {
-            w.git.submoduleUpdate().execute();
-            fail("Did not throw expected submodule update exception");
-        } catch (GitException e) {
-            assertThat(e.getMessage(), containsString("Command \"git submodule update modules/ntp\" returned status code 1"));
-        }
     }
 
     @Test
