+++ src/test/java/org/jenkinsci/plugins/gitclient/GitAPITestUpdateCliGit.java
@@ -1,11 +1,13 @@
 package org.jenkinsci.plugins.gitclient;
 
+import hudson.plugins.git.GitException;
 import org.junit.Test;
 
 import java.nio.file.Paths;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsString;
+import static org.junit.Assert.*;
 
 public abstract class GitAPITestUpdateCliGit extends GitAPITestUpdate{
 
@@ -29,6 +31,22 @@ public abstract class GitAPITestUpdateCliGit extends GitAPITestUpdate{
         String shallow = Paths.get(".git", "modules", "module", "1", "shallow").toString();
         assertFalse("shallow file existence: " + shallow, w.exists(shallow));
     }
+
+    @Test
+    public void testSubmoduleUpdateWithError() throws Exception {
+        w.git.clone_().url(localMirror()).execute();
+        w.git.checkout().ref("origin/tests/getSubmodules").execute();
+        w.rm("modules/ntp");
+        w.touch("modules/ntp", "file that interferes with ntp submodule folder");
+
+        try {
+            w.git.submoduleUpdate().execute();
+            fail("Did not throw expected submodule update exception");
+        } catch (GitException e) {
+            assertThat(e.getMessage(), containsString("Command \"git submodule update modules/ntp\" returned status code 1"));
+        }
+    }
+
 
 
 }
